requirements.txt
attrs==19.3.0
hypothesis==4.56.1
importlib-metadata==1.3.0
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pycryptodome==3.9.4
pyparsing==2.4.5
pytest==5.3.2
six==1.13.0
sortedcontainers==2.1.0
wcwidth==0.1.7
zipp==0.6.0
